# API_documentation.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: COURSE API
  description: This is the API Documentation of Course Recommender System of Team Sept-07
servers:
  - url: http://127.0.0.1:5001
...
paths:
...
  /api/course/{id}:
    description: End point to read a Course.
    get:
      tags:
        - Course
      summary: For Getting Course data
      description: Operation to get Course details.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              examples:
                Example 1:
                  value: |-
                    {
                        "data": {
                            "code": "CS001214",
                            "course_id": 1,
                            "course_type": "Diploma in Programming",
                            "description": "bla bla bla",
                            "duration": "12 weeks",
                            "is_active": true,
                            "level": "1",
                            "name": "Java",
                            "professor": "Someone"
                        }
                    }
        '404':
          description: Invalid Course Id
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Invalid Course id"}'
    put:
      tags:
        - Course
      summary: For updating Course data
      description: Operation to update Course Details.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Tools for Data Science
                code:
                  type: string
                  example: CS200123
                duration:
                  type: string
                  example: 12 Weeks
                professor:
                  type: string
                  example: M.K
                course_type:
                  type: string
                  example: Diploma in Data Science
                level:
                  type: string
                  example: '2'
                description:
                  type: string
                  example: Something bla bla
      responses:
        '200':
          description: Success
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Course Updated Successfully"}'
        '400':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Course Code is already taken"}'
        '404':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Invalid Course id"}'
    delete:
      tags:
        - Course
      summary: For deleting Course data
      description: Operation to delete a Course.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Course deleted Successfully"}'
        '404':
          description: Invalid Course Id
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Invalid Course id"}'
        '500':
          description: Invalid Course Id
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Something went wrong"}'
  /api/course:
    description: End point to create a Course.
    post:
      tags:
        - Course
      summary: For creating new Course
      description: Operation to create a Course.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Tools for Data Science
                code:
                  type: string
                  example: CS2100123
                duration:
                  type: string
                  example: 8 Weeks
                professor:
                  type: string
                  example: M.K
                course_type:
                  type: string
                  example: Diploma in Data Science
                level:
                  type: string
                  example: '2'
                description:
                  type: string
                  example: something something
      responses:
        '201':
          description: Success
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Course Added Successfully"}'
        '400':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Course Code is already taken"}'
                Example 2:
                  value: '{"message": "Course Name is already taken"}'
        '500':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Something went wrong"}'
  /api/courses/student/{user_id}:
    get:
      tags:
        - Course
      summary: For getting courses details for the student
      description: Operation to get User Course details.
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value:
                    Enrolled:
                      - - CS001214
                        - 12 weeks
                        - Someone
                        - 4
                        - 5
                        - 1
                        - Java
                    Unenrolled:
                      - - CS2001712
                        - 8 Weeks
                        - SLS
                        - 4
                        - 5
                        - 2
                        - System Commands
                    username: M.Ikram
        '404':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "user_id is invalid"}'
  /api/courses/student:
    post:
      tags:
        - Course
      summary: to get User Enrolled or unenrolled courses
      description: Operation to get User Enrolled or unenrolled courses details.
      requestBody:
        description: get User enrolled or unenrolled courses
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                  example: 1
                enrollment_type:
                  type: string
                  example: enrolled
      responses:
        '200':
          description: Success
          content:
            application/json:
              examples:
                Example 1:
                  value: |-
                    {
                        "data": [
                            {
                                "code": "CS001214",
                                "course_id": 1,
                                "course_type": "Diploma in Programming",
                                "description": "bla bla bla",
                                "duration": "12 weeks",
                                "is_active": true,
                                "level": "1",
                                "name": "Java",
                                "professor": "Someone"
                            }
                        ]
                    }
        '400':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Invalid enrollment_type"}'
...